feat: Overhaul ServiceCard component
Add three-dot dropdown menu
Add support to edit ServiceCard

+++ pages/api/service/[id].ts
@@ -8,46 +8,50 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
     method,
   } = req;
 
-  // let handledService = {};
+  if (!id) return;
 
-  switch (method) {
-    case 'DELETE':
-      if (!id) return;
-      try {
-        const handledService = await prisma.service.delete({
-          where: {
-            id: Array.isArray(id) ? parseInt(id[0]) : parseInt(id),
-          },
-        });
-        res.status(200).send(handledService);
-      } catch (error) {
-        res.status(500).send('Could not delete service');
-      }
+  // Get Service
+  if (method === 'GET') {
+    try {
+      const service = await prisma.service.findFirst({
+        where: {
+          id: Array.isArray(id) ? parseInt(id[0]) : parseInt(id),
+        },
+      });
+      return res.status(200).send(service);
+    } catch (error) {
+      return res.status(500).send('Could not delete service');
+    }
+    // Delete Service
+  } else if (method === 'DELETE') {
+    try {
+      const deletedService = await prisma.service.delete({
+        where: {
+          id: Array.isArray(id) ? parseInt(id[0]) : parseInt(id),
+        },
+      });
+      return res.status(200).send(deletedService);
+    } catch (error) {
+      return res.status(500).send('Could not delete service');
+    }
 
-      break;
-
-    case 'PUT':
-      if (!id) return;
-      try {
-        const data = req.body;
-        console.log(data);
-        const updatedService = await prisma.service.update({
-          data,
-          where: {
-            id: Array.isArray(id) ? parseInt(id[0]) : parseInt(id),
-          },
-        });
-
-        res.status(200).send(updatedService);
-      } catch (error) {
-        res.status(500).send('Could not edit service, try again');
-      }
-
-      break;
-
-    default:
-      res.setHeader('Allow', ['GET', 'PUT', 'DELETE']);
-      res.status(405).end(`Method ${method} Not Allowed`);
-      break;
+    // Update Service
+  } else if (method === 'PUT') {
+    try {
+      const data = req.body;
+      const updatedUser = await prisma.service.update({
+        data,
+        where: {
+          id: Array.isArray(id) ? parseInt(id[0]) : parseInt(id),
+        },
+      });
+      console.log(updatedUser)
+      return res.json(updatedUser);
+    } catch (error) {
+      return res.status(500).send('Could not edit service, try again');
+    }
   }
+
+  res.setHeader('Allow', ['GET', 'PUT', 'DELETE']);
+  res.status(405).end(`Method ${method} Not Allowed`);
 }
